refactor(motion-core): extract shared canvas color and pointer helpers
