+++ packages/motion-core/src/lib/components/fluid-image-reveal/component.json
@@ -20,6 +20,10 @@
 		{
 			"path": "../../utils/cn.ts",
 			"target": "utils"
+		},
+		{
+			"path": "../../helpers/fluid-pointer.ts",
+			"target": "helpers"
 		}
 	]
 }
