+++ packages/motion-core/src/lib/components/fluid-simulation/component.json
@@ -19,6 +19,14 @@
 		{
 			"path": "../../utils/cn.ts",
 			"target": "utils"
+		},
+		{
+			"path": "../../helpers/color.ts",
+			"target": "helpers"
+		},
+		{
+			"path": "../../helpers/fluid-pointer.ts",
+			"target": "helpers"
 		}
 	]
 }
