+++ packages/motion-core/src/lib/components/lava-lamp/LavaLampScene.svelte
@@ -10,6 +10,7 @@
 		Vec3,
 		Vec4,
 	} from "ogl";
+	import { toLinearRgb } from "../../helpers/color";
 
 	interface Props {
 		/**
@@ -63,80 +64,6 @@
 		uRadius: { value: number };
 		uSmoothness: { value: number };
 	}>();
-
-	const clamp01 = (value: number) => Math.min(1, Math.max(0, value));
-	const srgbToLinear = (value: number) =>
-		value <= 0.04045 ? value / 12.92 : Math.pow((value + 0.055) / 1.055, 2.4);
-
-	const parseHexColor = (value: string): [number, number, number] | null => {
-		const hex = value.replace("#", "").trim();
-		if (hex.length === 3 || hex.length === 4) {
-			const r = Number.parseInt(hex[0] + hex[0], 16);
-			const g = Number.parseInt(hex[1] + hex[1], 16);
-			const b = Number.parseInt(hex[2] + hex[2], 16);
-			return [r / 255, g / 255, b / 255];
-		}
-		if (hex.length === 6 || hex.length === 8) {
-			const r = Number.parseInt(hex.slice(0, 2), 16);
-			const g = Number.parseInt(hex.slice(2, 4), 16);
-			const b = Number.parseInt(hex.slice(4, 6), 16);
-			return [r / 255, g / 255, b / 255];
-		}
-		return null;
-	};
-
-	let cssColorContext: CanvasRenderingContext2D | null | undefined;
-	const parseCssColor = (value: string): [number, number, number] | null => {
-		if (typeof document === "undefined") return null;
-		if (cssColorContext === undefined) {
-			const parserCanvas = document.createElement("canvas");
-			parserCanvas.width = 1;
-			parserCanvas.height = 1;
-			cssColorContext = parserCanvas.getContext("2d");
-		}
-		if (!cssColorContext) return null;
-
-		cssColorContext.fillStyle = "#000000";
-		cssColorContext.fillStyle = value;
-		const normalized = cssColorContext.fillStyle;
-
-		if (normalized.startsWith("#")) {
-			return parseHexColor(normalized);
-		}
-
-		const match = normalized.match(/rgba?\(([^)]+)\)/i);
-		if (!match) return null;
-		const parts = match[1]
-			.split(",")
-			.map((part) => Number.parseFloat(part.trim()))
-			.filter((part) => Number.isFinite(part));
-		if (parts.length < 3) return null;
-		const scale = Math.max(parts[0], parts[1], parts[2]) > 1 ? 255 : 1;
-		return [
-			clamp01(parts[0] / scale),
-			clamp01(parts[1] / scale),
-			clamp01(parts[2] / scale),
-		];
-	};
-
-	const toRgb = (
-		value: string,
-		fallback: [number, number, number],
-	): [number, number, number] => {
-		const trimmed = value.trim();
-		const parsed = trimmed.startsWith("#")
-			? parseHexColor(trimmed)
-			: parseCssColor(trimmed);
-		return parsed ?? fallback;
-	};
-
-	const toLinearRgb = (
-		value: string,
-		fallback: [number, number, number],
-	): [number, number, number] => {
-		const [r, g, b] = toRgb(value, fallback);
-		return [srgbToLinear(r), srgbToLinear(g), srgbToLinear(b)];
-	};
 
 	const applyColor = (
 		target: Vec3,
