+++ packages/motion-core/src/lib/components/rubiks-cube/component.json
@@ -19,6 +19,10 @@
 		{
 			"path": "../../utils/cn.ts",
 			"target": "utils"
+		},
+		{
+			"path": "../../helpers/color.ts",
+			"target": "helpers"
 		}
 	]
 }
