+++ packages/motion-core/src/lib/components/specular-band/component.json
@@ -20,6 +20,10 @@
 		{
 			"path": "../../utils/cn.ts",
 			"target": "utils"
+		},
+		{
+			"path": "../../helpers/color.ts",
+			"target": "helpers"
 		}
 	]
 }
